frontend/requirements.txt
streamlit
pydeck
pandas
requests
python-dotenv
